added lighting / fix skybox later

+++ include/rg/Camera.h
@@ -83,7 +83,7 @@ public:
     }
 
     void ProcessMouseScroll(float yoffset) {
-        Zoom -= yoffset;
+        Zoom -= (float)yoffset;
         if (Zoom < 1.0f) {
             Zoom = 1.0f;
         }
